docker_images/dlm-dti/requirements.txt
PyYAML==6.0.1
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
pytorch-lightning==1.8.4
tqdm==4.65.0
scikit-learn==1.0.2
numpy==1.21.6
pandas==1.3.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
notebook==6.5.4
notebook_shim==0.2.3
transformers==4.30.2
